Improve accessibility with aria-labelledby
From the W3C-ARIA Spec [http://www.w3.org/TR/wai-aria/roles#region]:

```
Authors SHOULD ensure that a region has a heading referenced by aria-labelledby.
This heading is provided by an instance of the standard host language heading
element or an instance of an element with role heading that contains the heading
text.
```

In reality what this does is means that when the whole region is navigated to
using a screen reader the screen reader will read out the heading and
-importantly- allow the user to skip the whole region if they want to. Without
the aria-labelledby this doesn't work.